+++ app/Models/Project.php
@@ -40,6 +40,11 @@ class Project extends Model {
         return $this->hasMany(User::class);
     }
 
+    public function users()
+    {
+        return $this->belongsToMany(User::class);
+    }
+
     public function tasks()
     {
         return $this->hasMany(Task::class);