+++ app/Models/User.php
@@ -107,6 +107,11 @@ class User extends Authenticatable {
         return $this->hasMany(Task::class);
     }
 
+    public function projects()
+    {
+        return $this->belongsToMany(Project::class);
+    }
+
     public static function toOptions($where = NULL)
     {
         $query = self::query();