+++ resources/views/project/partials/form.blade.php
@@ -27,6 +27,28 @@
             @enderror
         </div>
 
+        <div class="mb-3">
+            <label class="form-label">
+                {{ __('Users') }}
+            </label>
+
+            @if($userOptions->isEmpty())
+                <x-no-records-found/>
+            @endif
+
+            @foreach($userOptions as $userOption)
+                <div class="form-check form-switch mb-2">
+                    <input class="form-check-input" type="checkbox" role="switch"
+                           id="project-user-{{$userOption->value}}" name="users[]" value="{{$userOption->value}}"
+                           @if($project->users->contains($userOption->value)) checked @endif>
+                    <label class="form-check-label small" for="project-user-{{$userOption->value}}">
+                        {{$userOption->label}}
+                    </label>
+                </div>
+            @endforeach
+
+        </div>
+
     </div>
 
 </div>